cloud_run/e2e_tests/requirements.txt
google-auth==2.6.3
requests==2.27.1
google-cloud-bigquery==2.34.0
google-cloud-storage==2.1.0
google-cloud-secret-manager==1.0.1
pyarrow==7.0.0
pandas==1.3.5
protobuf==3.20.*
